Add feature: read only keys.

+++ KeyStore.c
@@ -33,6 +33,7 @@ void delete_element(unsigned int index)
     }
 }
 
+
 static
 unsigned int find_element(const unsigned int app_id, const char name [KEY_NAME_SIZE])
 {
@@ -45,8 +46,8 @@ unsigned int find_element(const unsigned int app_id, const char name [KEY_NAME_S
     {
         if (!key_store.element_store[k].admin.is_free)
         {
-            if (app_id == key_store.element_store[k].admin.app_id && 
-		0 == memcmp(name, (void *) key_store.element_store[k].key.name, KEY_NAME_SIZE))
+            if (app_id == key_store.element_store[k].admin.app_id &&
+                0 == memcmp(name, (void *) key_store.element_store[k].key.name, KEY_NAME_SIZE))
             {
                 return k;
             }
@@ -83,11 +84,53 @@ void key_store_init(void)
 }
 
 
+unsigned int key_store_init_with_read_only_keys(unsigned int const *app_ids, key_record_t const *keys, unsigned int nr_keys)
+{
+    if (nr_keys > NR_ELEMENTS)
+    {
+        return -1;
+    }
+
+    key_store.free_slots = NR_ELEMENTS - nr_keys;
+    for (unsigned int k = 0; k < NR_ELEMENTS; k++)
+    {
+        key_store.element_store[k].admin.is_free = 1;
+    }
+
+    for (unsigned int k = 0; k < nr_keys; k++)
+    {
+        if (find_element (app_ids[k], keys[k].name) < NR_ELEMENTS)
+        {
+            return -1;
+        }
+
+        key_store.element_store[k].admin.is_free = 0;
+        key_store.element_store[k].admin.app_id = app_ids[k];
+        key_store.element_store[k].key.read_only = 1;
+        memcpy(key_store.element_store[k].key.name, keys[k].name, KEY_NAME_SIZE);
+        memcpy(key_store.element_store[k].key.data, keys[k].data, KEY_DATA_SIZE);
+    }
+
+    for (unsigned int k = nr_keys; k < NR_ELEMENTS; k++)
+    {
+        key_store.element_store[k].admin.app_id = 0;
+        key_store.element_store[k].key.read_only = 0;
+        reset_element_key(k);
+    }
+
+    return 0;
+}
+
+
 void key_store_wipe(void)
 {
     for (unsigned int k = 0; k < NR_ELEMENTS; k++)
     {
-        delete_element(k);
+    if (!key_store.element_store[k].admin.is_free &&
+        !key_store.element_store[k].key.read_only)
+        {
+            delete_element(k);
+        }
     }
 }
 
@@ -128,6 +171,7 @@ key_store_result_t key_store_add(unsigned int app_id, key_record_t const *key)
     key_store.element_store[result.index].admin.app_id = app_id;
     key_store.element_store[result.index].admin.is_free = 0;
 
+    key_store.element_store[result.index].key.read_only = 0;
     memcpy(key_store.element_store[result.index].key.name, key->name, KEY_NAME_SIZE);
     memcpy(key_store.element_store[result.index].key.data, key->data, KEY_DATA_SIZE);
 
@@ -164,6 +208,7 @@ key_store_result_t key_store_get(unsigned int app_id, const char name [KEY_NAME_
 
     memcpy(key->name, key_store.element_store[result.index].key.name, KEY_NAME_SIZE);
     memcpy(key->data, key_store.element_store[result.index].key.data, KEY_DATA_SIZE);
+    key->read_only = key_store.element_store[result.index].key.read_only;
 
     result.error = 0;
     return result;
@@ -198,6 +243,7 @@ unsigned int key_store_get_by_index(unsigned int app_id, unsigned int index, key
 
     memcpy(key->name, key_store.element_store[index].key.name, KEY_NAME_SIZE);
     memcpy(key->data, key_store.element_store[index].key.data, KEY_DATA_SIZE);
+    key->read_only = key_store.element_store[index].key.read_only;
 
     return 0;
 }
@@ -221,7 +267,12 @@ unsigned int key_store_delete(unsigned int app_id, const char name [KEY_NAME_SIZ
         return -1;
     }
 
-    delete_element(element_index);    
+    if (key_store.element_store[element_index].key.read_only)
+    {
+        return -1;
+    }
+
+    delete_element(element_index);
 
     return 0;
 }
